add checkbox and move loginform to own component
